test(BA-7154): move the client registry fixtures into conftest
The two registries were declared in the test module while every other fixture
for this directory already lived in conftest, so the arrange was split across
two files for no reason. Directories that have a conftest keep these there —
13 do, and the handful that inline them have no conftest to use.

The test module is now tests only.
