+++ tests/component/app_config/conftest.py
@@ -6,7 +6,11 @@ from unittest.mock import MagicMock
 
 import pytest
 import sqlalchemy as sa
+import yarl
 
+from ai.backend.client.v2.auth import HMACAuth, NoAuth
+from ai.backend.client.v2.config import ClientConfig
+from ai.backend.client.v2.v2_registry import V2ClientRegistry
 from ai.backend.common.data.app_config.types import AppConfigScopeType
 from ai.backend.common.identifier.app_config import AppConfigScopeID
 from ai.backend.manager.api.adapters.app_config.adapter import AppConfigAdapter
@@ -26,7 +30,7 @@ from ai.backend.manager.services.app_config.processors import AppConfigProcessor
 from ai.backend.manager.services.app_config.service import AppConfigService
 
 if TYPE_CHECKING:
-    from tests.component.conftest import UserFixtureData
+    from tests.component.conftest import ServerInfo, UserFixtureData
 
 
 @pytest.fixture()
@@ -106,3 +110,31 @@ async def merged_fragments(
                 AppConfigDefinitionRow.config_name.in_(config_names)
             )
         )
+
+
+@pytest.fixture()
+async def user_v2_registry(
+    server: ServerInfo,
+    regular_user_fixture: UserFixtureData,
+) -> AsyncIterator[V2ClientRegistry]:
+    registry = await V2ClientRegistry.create(
+        ClientConfig(endpoint=yarl.URL(server.url)),
+        HMACAuth(
+            access_key=regular_user_fixture.keypair.access_key,
+            secret_key=regular_user_fixture.keypair.secret_key,
+        ),
+    )
+    try:
+        yield registry
+    finally:
+        await registry.close()
+
+
+@pytest.fixture()
+async def anonymous_v2_registry(server: ServerInfo) -> AsyncIterator[V2ClientRegistry]:
+    """A caller that holds no credentials at all — the pre-login client."""
+    registry = await V2ClientRegistry.create(ClientConfig(endpoint=yarl.URL(server.url)), NoAuth())
+    try:
+        yield registry
+    finally:
+        await registry.close()
